Fix warnings

+++ src/fft.cpp
@@ -22,16 +22,16 @@ void Fft::init(unsigned width, unsigned height, unsigned num_of_feats, unsigned 
 void Fft::set_window(const MatDynMem &window)
 {
     assert(window.dims == 2);
-    assert(window.size().width == m_width);
-    assert(window.size().height == m_height);
+    assert(window.size().width == int(m_width));
+    assert(window.size().height == int(m_height));
     (void)window;
 }
 
 void Fft::forward(const MatDynMem &real_input, ComplexMat &complex_result)
 {
     assert(real_input.dims == 2);
-    assert(real_input.size().width == m_width);
-    assert(real_input.size().height == m_height);
+    assert(real_input.size().width == int(m_width));
+    assert(real_input.size().height == int(m_height));
     (void)real_input;
     (void)complex_result;
 }
@@ -40,12 +40,12 @@ void Fft::forward_window(MatDynMem &patch_feats, ComplexMat &complex_result, Mat
 {
         assert(patch_feats.dims == 3);
 #ifndef BIG_BATCH
-        assert(patch_feats.size[0] == m_num_of_feats);
+        assert(patch_feats.size[0] == int(m_num_of_feats));
 #else
-        assert(patch_feats.size[0] == m_num_of_feats * m_num_of_scales);
+        assert(patch_feats.size[0] == int(m_num_of_feats * m_num_of_scales));
 #endif
-        assert(patch_feats.size[1] == m_height);
-        assert(patch_feats.size[2] == m_width);
+        assert(patch_feats.size[1] == int(m_height));
+        assert(patch_feats.size[2] == int(m_width));
 
         assert(tmp.dims == patch_feats.dims);
         assert(tmp.size[0] == patch_feats.size[0]);
@@ -61,12 +61,12 @@ void Fft::inverse(ComplexMat &complex_input, MatDynMem &real_result)
 {
     assert(real_result.dims == 3);
 #ifndef BIG_BATCH
-    assert(real_result.size[0] == m_num_of_feats);
+    assert(real_result.size[0] == int(m_num_of_feats));
 #else
-    assert(real_result.size[0] == m_num_of_feats * m_num_of_scales);
+    assert(real_result.size[0] == int(m_num_of_feats * m_num_of_scales));
 #endif
-    assert(real_result.size[1] == m_height);
-    assert(real_result.size[2] == m_width);
+    assert(real_result.size[1] == int(m_height));
+    assert(real_result.size[2] == int(m_width));
 
     (void)complex_input;
     (void)real_result;
